requirements.txt
scipy==1.5.4
numpy==1.21.6
catboost==1.1.1
optuna==3.0.6
scikit-optimize==0.9.0
imbalanced-learn==0.8.1
category-encoders==2.6.0
xverse==1.0.5 shap==0.41.0
explainerdashboard==0.3.8.2
pmdarima==2.0.4
seaborn==0.11.2
spark-sklearn==0.2.1
plotly==5.13.1
pandas-summary==0.2.0
google-cloud-bigquery==3.4.1
google-cloud-bigquery-storage==2.13.2
google-cloud-secret-manager==2.10.0
google-auth==2.25.2
pandas-gbq==0.18.1
google-cloud-storage==2.7.0
gcsfs==0.8.0
pyarrow==6.0.1
rank_bm25
matplotlib==3.3.4
scikit-learn==0.24.2
statsmodels==0.12.2
db-dtypes==1.0.2
psycopg2
sqlalchemy-bigquery
joblib==1.3.2
lightgbm==3.3.3